Move model content to new activeproject, archivedproject, legacy, unpublishedproject modules
