+++ physionet-django/project/modelcomponents/coreproject.py
@@ -3,6 +3,10 @@ import uuid
 from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
 
+from project.modelcomponents.activeproject import ActiveProject
+from project.modelcomponents.archivedproject import ArchivedProject
+from project.modelcomponents.publishedproject import PublishedProject
+
 
 class CoreProject(models.Model):
     """
@@ -64,3 +68,13 @@ class ProgrammingLanguage(models.Model):
 
     def __str__(self):
         return self.name
+
+
+def exists_project_slug(slug):
+    """
+    Whether the slug has been taken by an existing project of any
+    kind.
+    """
+    return bool(ActiveProject.objects.filter(slug=slug)
+            or ArchivedProject.objects.filter(slug=slug)
+            or PublishedProject.objects.filter(slug=slug))
